fix(Contour): Fix value when first contour is added
When there are no current contours and the first was added, it was adding null as the new value.
This was causing the contouring to fail, so I changed it to add the midpoint value of the current
array instead of the value of the last contour.  This new equation always returns a number and
should work better.